Refactor packages and files
