+++ app/src/main/kotlin/com/agilefreaks/freaks_catalog/FreaksCatalogApp.kt
@@ -1,6 +1,6 @@
 package com.agilefreaks.freaks_catalog
 
-import com.agilefreaks.freaks_catalog.features.freaks.detailsFragmentModule
+import com.agilefreaks.freaks_catalog.features.freaks.details.detailsFragmentModule
 import com.agilefreaks.freaks_catalog.features.freaks.freaksFragmentModule
 import com.google.android.play.core.splitcompat.SplitCompatApplication
 import org.koin.android.ext.koin.androidContext
