+++ app/src/main/res/values/colors.xml
@@ -2,5 +2,5 @@
 <resources>
     <color name="gray">#808080</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="primary_blue">#317FF9</color>
+    <color name="primary_blue">#105FBF</color>
 </resources>
