+++ features/freaks/src/main/kotlin/com/agilefreaks/freaks_catalog/features/freaks/FreaksDataSource.kt
@@ -1,6 +1,5 @@
 package com.agilefreaks.freaks_catalog.features.freaks
 
-import com.agilefreaks.freaks_catalog.features.freaks.model.apolloClient
 import com.apollographql.apollo.coroutines.await
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
