+++ features/freaks/src/main/kotlin/com/agilefreaks/freaks_catalog/features/freaks/FreaksFragmentModule.kt
@@ -1,7 +1,6 @@
 package com.agilefreaks.freaks_catalog.features.freaks
 
-import com.agilefreaks.freaks_catalog.features.freaks.model.FilterViewModel
-import com.agilefreaks.freaks_catalog.features.freaks.model.FreaksViewModel
+import com.agilefreaks.freaks_catalog.features.freaks.filter.FilterViewModel
 import com.agilefreaks.freaks_catalog.features.freaks.repository.FreaksRepository
 import com.agilefreaks.freaks_catalog.features.freaks.repository.FreaksRepositoryImpl
 import org.koin.dsl.module
