+++ features/freaks/src/main/res/drawable/details_image_background.xml
