+++ features/freaks/src/main/res/values/colors.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="variant_blue">#105FBF</color>
+    <color name="variant_blue">#E6105FBF</color>
 </resources>
