+++ features/freaks/src/test/kotlin/com/agilefreaks/freaks_catalog/features/freaks/list/FreaksRepositoryTest.kt
@@ -2,7 +2,7 @@ package com.agilefreaks.freaks_catalog.features.freaks.list
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.agilefreaks.freaks_catalog.features.freaks.CoroutineRule
-import com.agilefreaks.freaks_catalog.features.freaks.Freak
+import com.agilefreaks.freaks_catalog.features.freaks.model.Freak
 import com.agilefreaks.freaks_catalog.features.freaks.FreaksDataSource
 import com.agilefreaks.freaks_catalog.features.freaks.FreaksListQuery
 import com.agilefreaks.freaks_catalog.features.freaks.repository.FreaksRepositoryImpl
