+++ features/freaks/src/test/kotlin/com/agilefreaks/freaks_catalog/features/freaks/model/FilterViewModelTest.kt
@@ -1,7 +1,8 @@
-package com.agilefreaks.freaks_catalog.features.freaks
+package com.agilefreaks.freaks_catalog.features.freaks.model
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
-import com.agilefreaks.freaks_catalog.features.freaks.model.FilterViewModel
+import com.agilefreaks.freaks_catalog.features.freaks.CoroutineRule
+import com.agilefreaks.freaks_catalog.features.freaks.filter.FilterViewModel
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Rule
@@ -19,8 +20,8 @@ class FilterViewModelTest {
     @Test
     fun `init will load skills with all the proper skill names`() {
         val viewModel = FilterViewModel()
-        val skillNames = viewModel.skills.value?.map { it.skillName }
-        assertThat(skillNames).containsExactly("Android", "Kotlin", "Other Skill", "iOS", "Ruby", "QA")
+        val skillNames = viewModel.skills.value?.map { it.name }
+        assertThat(skillNames).containsExactly("Android", "Kotlin", "Other Skill", "QA", "Ruby", "iOS")
     }
 
     @ExperimentalCoroutinesApi
@@ -35,8 +36,8 @@ class FilterViewModelTest {
     @Test
     fun `init will load projects`() {
         val viewModel = FilterViewModel()
-        val projectNames = viewModel.projects.value?.map { it.skillName }
-        assertThat(projectNames).containsExactly("Freaks Catalog", "Proj2", "Tutorial", "Altceva")
+        val projectNames = viewModel.projects.value?.map { it.name }
+        assertThat(projectNames).containsExactly("EPIX", "Freaks Catalog", "New Project", "reAsig")
     }
 
     @Test
